+++ lib/pages/mainPage.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:libacflutter/Constants.dart';
@@ -79,6 +81,15 @@ class _AccountPage extends State<AccountPage> {
 
   @override
   Widget build(BuildContext context) {
+    if (ms.currentUser.ID == UUID_ZERO) {
+      Timer(Duration(seconds: 5), () async {
+        // Obtain the current user!
+        var reply = await NetworkInterface.getUser(ms.username);
+        if (reply.success) {
+          ms.currentUser = reply.data!;
+        }
+      });
+    }
     return Scaffold(
       appBar: AppBar(
         title: Text("Switchboard"),