+++ lib/pages/settings_categories/account.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:libacflutter/Constants.dart';
+import 'package:libacflutter/Prompt.dart';
 import 'package:switchboard/dart/MemoryState.dart';
 import 'package:switchboard/dart/storage.dart';
+import 'package:switchboard/dart/users.dart';
 import 'package:switchboard/globalHelpers.dart';
 
 class AccountSettings extends StatefulWidget {
@@ -35,6 +38,106 @@ class _acct extends State<AccountSettings> {
         child: SingleChildScrollView(
           child: Column(
             children: [
+              Center(child: Text("Info", style: TextStyle(fontSize: 20))),
+              Divider(),
+              Card(
+                elevation: 8,
+                child: Padding(
+                  padding: EdgeInsetsGeometry.all(8),
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.start,
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      ListTile(
+                        title: Text("Username"),
+                        subtitle: Text(MemoryState.A.username),
+                        leading: Icon(Icons.person),
+                      ),
+                      SizedBox(height: 8),
+                      ListTile(
+                        title: Text("User ID"),
+                        subtitle: Text(
+                          "${MemoryState.A.currentUser.ID}\nTap to copy | Needed for Wiping or Deleting account",
+                        ),
+                        leading: Icon(Icons.perm_identity),
+                        onTap: () async {
+                          await Clipboard.setData(
+                            ClipboardData(text: MemoryState.A.currentUser.ID),
+                          );
+                        },
+                        shape: RoundedRectangleBorder(
+                          borderRadius: BorderRadiusGeometry.circular(8),
+                        ),
+                      ),
+                      SizedBox(height: 8),
+                      ListTile(
+                        title: Text("Account Level"),
+                        subtitle: Text(
+                          getAccountLevel(
+                            MemoryState.A.currentUser.AccountLevel,
+                          ),
+                        ),
+                        leading: Icon(Icons.key_sharp),
+                      ),
+                      SizedBox(height: 8),
+                      FutureBuilder(
+                        future: NetworkInterface.requestAltersList(null),
+                        builder: (bldr3, snp3) {
+                          if (!snp3.hasData) {
+                            return CircularProgressIndicator();
+                          } else {
+                            MemoryState.A.currentUser.AlterCount =
+                                snp3.data!.alters.length;
+
+                            return ListTile(
+                              title: Text("Alter Count"),
+                              subtitle: Text("${snp3.data!.alters.length}"),
+                              leading: Icon(Icons.numbers),
+                            );
+                          }
+                        },
+                      ),
+                      SizedBox(height: 8),
+                      ListTile(
+                        title: Text("Friend Count"),
+                        subtitle: Text("0"),
+                        leading: Icon(Icons.numbers),
+                      ),
+                      SizedBox(height: 8),
+                      if ((MemoryState.A.currentUser.AccountLevel &
+                                  USER_LEVEL_TESTER) ==
+                              USER_LEVEL_TESTER ||
+                          (MemoryState.A.currentUser.AccountLevel &
+                                  USER_LEVEL_APP_STORE) ==
+                              USER_LEVEL_APP_STORE)
+                        ListTile(
+                          title: Text("Password cannot be changed"),
+                          subtitle: Text(
+                            "Your account level is either that of a tester, or a App Store employee, you cannot change security features related to this limited account",
+                          ),
+                          leading: Icon(Icons.warning, color: Colors.yellow),
+                          tileColor: LibACFlutterConstants.TITLEBAR_COLOR,
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadiusGeometry.circular(8),
+                          ),
+                        ),
+                      if (!((MemoryState.A.currentUser.AccountLevel &
+                                  USER_LEVEL_TESTER) ==
+                              USER_LEVEL_TESTER) &&
+                          !((MemoryState.A.currentUser.AccountLevel &
+                                  USER_LEVEL_APP_STORE) ==
+                              USER_LEVEL_APP_STORE))
+                        ListTile(
+                          title: Text("Security Options"),
+                          subtitle: Text("Password, Email..."),
+                          leading: Icon(Icons.security),
+                          trailing: Icon(Icons.forward),
+                        ),
+                    ],
+                  ),
+                ),
+              ),
+              SizedBox(height: 25),
               Center(child: Text("Options", style: TextStyle(fontSize: 20))),
               Divider(),
               Card(
@@ -218,31 +321,63 @@ class _acct extends State<AccountSettings> {
                                     isDestructiveAction: true,
                                     child: Text("Yes"),
                                     onPressed: () async {
-                                      var reply =
-                                          await NetworkInterface.wipeAccount();
-                                      if (reply.success) {
-                                        ScaffoldMessenger.of(
-                                          context,
-                                        ).showSnackBar(
-                                          SnackBar(
-                                            content: Text(
-                                              "Your request succeeded. Account Erased!",
-                                            ),
-                                          ),
-                                        );
-                                      } else {
-                                        ScaffoldMessenger.of(
-                                          context,
-                                        ).showSnackBar(
-                                          SnackBar(
-                                            content: Text(
-                                              "Your request failed. Account was not erased.\nID: ${reply.id}\nReason: ${reply.reason}",
-                                            ),
-                                          ),
-                                        );
-                                      }
+                                      await showDialog(
+                                        context: context,
+                                        builder: (bldr2) {
+                                          return InputPrompt(
+                                            title: "Please Confirm",
+                                            prompt:
+                                                "User UUID Required. Enter it here to confirm.",
+                                            type: InputPromptType.Text,
+                                            successAction: (p0) async {
+                                              if (p0 !=
+                                                  MemoryState
+                                                      .A
+                                                      .currentUser
+                                                      .ID) {
+                                                // ABORT!!
+                                                ScaffoldMessenger.of(
+                                                  context,
+                                                ).showSnackBar(
+                                                  SnackBar(
+                                                    content: Text(
+                                                      "Account not wiped, confirmation challenge failed",
+                                                    ),
+                                                  ),
+                                                );
+                                                return;
+                                              }
+                                              var reply =
+                                                  await NetworkInterface.wipeAccount();
+                                              if (reply.success) {
+                                                ScaffoldMessenger.of(
+                                                  context,
+                                                ).showSnackBar(
+                                                  SnackBar(
+                                                    content: Text(
+                                                      "Your request succeeded. Account Erased!",
+                                                    ),
+                                                  ),
+                                                );
+                                              } else {
+                                                ScaffoldMessenger.of(
+                                                  context,
+                                                ).showSnackBar(
+                                                  SnackBar(
+                                                    content: Text(
+                                                      "Your request failed. Account was not erased.\nID: ${reply.id}\nReason: ${reply.reason}",
+                                                    ),
+                                                  ),
+                                                );
+                                              }
 
-                                      setState(() {});
+                                              setState(() {});
+                                              Navigator.pop(context);
+                                            },
+                                          );
+                                        },
+                                      );
+
                                       Navigator.pop(context);
                                     },
                                   ),